Change /csrf endpoint to send cookie
When a GET request is sent to the /csrf endpoint, the response
sets a cookie containing a secret value used for CSRF prevention.
A custom header is also set that contains a token derived from
this secret value. This token, along with the secret value, can
later be used to verify that any POST requests sent to the server
were authorized.