Send emails on contact form submission
The POST handler function that is called when the contact form
is submitted was modified to call a function that will send an
acknowledgement email to the user every time they submit the
contact form. As of yet, this function doesn't actually send any
email (it still needs to be written).

The handler also calls a function that should send an email to
the site owner letting them know that someone has requested to
contact them. Again, this function doesn't actually do anything
right now (will be defined later).