Ensure code complies with TypeScript rules
Several changes were made to many files to ensure that the
TypeScript compiler doesn't raise any errors during compilation.

Also, a build script was written. So far, it doesn't do anything
more than print "Hello, World!" to the console. This is strictly
to sanity-test that the script is actually run after compilation.