Move database connection logic to database.ts file
The bin/build.ts file previously performed some error checking to
ensure that the environment variables needed to connect to the
application's database were provided. Furthermore, the same file
would create the object passed to the mysql2.createConnection
function to connect to the database.

This logic has all been moved to the src/database.ts file. It
makes more sense for these functions to be defined in the file
that contains all the functions needed to communicate with the
database.