Write function to create database record
When a user submits a request to be contacted, a function is
called that enters a new record in the database containing their
contact info. This function previously did nothing, but now it
actually connects to the database and runs the appropriate INSERT
statement to create a new database record.