Create function for retrieving env variables
A new utility function was created that retrieves environment
variables. It can retrieve these variables from a .env file
specified in a command-line argument, or it can check if they
are already defined in the environment.

A second function was defined that ensures the required
environment variables are set. This prevents the app from crashing
unexpectedly at some point. Both of these functions will be called
when the build command is run and when the server is started to
ensure that the required environment variables are set before
anything else is done.