Refactor app to use new utility functions
In the last commit, two new utility functions were defined: One
that retrieves the environment variables needed to build and
serve the app, and another to validate the values of these
environment variables.

A few changes were made to the latter of these two functions so
that it better validates the environment variables. Furthermore,
several other files were refactored to now use these new
functions. For instance, the bin/build.ts file that defines the
actions performed when the build command is run now uses these
two functions instead of retrieving and validating the
environment variables itself.

Also, tslib was installed. This is because, every time the app's
tests were run, Jest would complain that tslib was not installed
and the AWS SDK depends on it.