Refactor application files
Several of the files were placed in subdirectories of the /src
directory. This ensures that everything is better
compartmentalized and easier to keep track of, instead of having
a flat directory structure in which every file is located in the
same directory.

Furthermore, the utils.ts file was broken up into several differnt
files within a /src/utils directory. The previous utils.ts file
was getting too large and difficult to manage.