Write middleware to validate CSRF cookie values
A new middleware function was written that validates the value
of the csrftoken cookie and the X-CSRF-Token header in POST
requests to the /contact endpoint. The app was configured to
use this new middleware.

Some of the tests that are run against the app had to be adjusted
to accommodate this new middleware.