Write functions to save emails on disk
Several functions were written that can compose emails and store
them as files on the local filesystem. If the EMAIL_DRIVER
environment variable is set to 'local', these functions are used.
Any time a user sends a POST request to the /contact endpoint,
an acknowledgement email is saved as a file on disk and an email
addressed to the site's admin is saved on disk.

The functions that send real emails using AWS Simple Email
Service still need to be written. At that point, version 1.0 of
the app will be complete.